Rename src package
